requirements.txt
yt-dlp
requests
Pillow
mutagen